# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/services/{id}": {
      "get": {
        "operationId": "ServicesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Buscar um Services",
        "tags": [
          "Services"
        ]
      },
      "put": {
        "operationId": "ServicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Atualizar Services",
        "tags": [
          "Services"
        ]
      },
      "delete": {
        "operationId": "ServicesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Remover Services",
        "tags": [
          "Services"
        ]
      }
    },
    "/quotes": {
      "post": {
        "operationId": "QuotesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Quotes criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Criar Quotes",
        "tags": [
          "Quotes"
        ]
      },
      "get": {
        "operationId": "QuotesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Listar todos Quotes",
        "tags": [
          "Quotes"
        ]
      }
    },
    "/quotes/{id}": {
      "get": {
        "operationId": "QuotesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Buscar um Quotes",
        "tags": [
          "Quotes"
        ]
      },
      "put": {
        "operationId": "QuotesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Atualizar Quotes",
        "tags": [
          "Quotes"
        ]
      },
      "delete": {
        "operationId": "QuotesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Remover Quotes",
        "tags": [
          "Quotes"
        ]
      }
    },
    "/quotes/{id}/sign": {
      "post": {
        "operationId": "QuotesController_saveSignature",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Quotes criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation saveSignature",
        "tags": [
          "Quotes"
        ]
      }
    },
    "/quotes/{id}/pdf": {
      "get": {
        "operationId": "QuotesController_getPdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation getPdf",
        "tags": [
          "Quotes"
        ]
      }
    },
    "/public/quotes/{id}": {
      "get": {
        "operationId": "QuotesPublicController_findPublicQuote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation findPublicQuote",
        "tags": [
          "Quotes-public"
        ]
      }
    },
    "/public/quotes/{id}/sign": {
      "post": {
        "operationId": "QuotesPublicController_savePublicSignature",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Quotes-public criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation savePublicSignature",
        "tags": [
          "Quotes-public"
        ]
      }
    },
...
    "/service-orders/{id}": {
      "get": {
        "operationId": "ServiceOrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Buscar um Service-orders",
        "tags": [
          "Service-orders"
        ]
      },
      "put": {
        "operationId": "ServiceOrdersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Atualizar Service-orders",
        "tags": [
          "Service-orders"
        ]
      }
    },
...
    "/public/service-orders/{id}": {
      "get": {
        "operationId": "ServiceOrdersPublicController_findPublicOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation findPublicOrder",
        "tags": [
          "Service-orders-public"
        ]
      }
    },
    "/public/service-orders/{id}/rate": {
      "post": {
        "operationId": "ServiceOrdersPublicController_saveClientRating",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Service-orders-public criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation saveClientRating",
        "tags": [
          "Service-orders-public"
        ]
      }
    },
...
  },
  "info": {
    "title": "Click Marido API",
    "description": "API documentation for Click Marido ERP/CRM",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "UpdateServiceDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "description": "Campo category",
            "example": "exemplo"
          },
          "name": {
            "type": "string",
            "description": "Campo name",
            "example": "exemplo"
          },
          "description": {
            "type": "string",
            "description": "Campo description",
            "example": "exemplo"
          },
          "value": {
            "type": "number",
            "description": "Campo value",
            "example": 1
          },
          "averageTime": {
            "type": "number",
            "description": "Campo averageTime",
            "example": 1
          },
          "complexity": {
            "type": "string",
            "description": "Campo complexity",
            "example": "exemplo"
          },
          "warranty": {
            "type": "number",
            "description": "Campo warranty",
            "example": 1
          },
          "specialty": {
            "type": "string",
            "description": "Campo specialty",
            "example": "exemplo"
          },
          "active": {
            "type": "boolean",
            "description": "Campo active",
            "example": true
          }
        }
      },
      "CreateQuoteDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "description": "Campo clientId",
            "example": "exemplo"
          },
          "discount": {
            "type": "number",
            "description": "Campo discount",
            "example": 1
          },
          "travelFee": {
            "type": "number",
            "description": "Campo travelFee",
            "example": 1
          },
          "materials": {
            "description": "Campo materials",
            "example": "exemplo",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "description": "Campo status",
            "example": "exemplo"
          },
          "services": {
            "description": "Campo services",
            "example": "exemplo",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "clientId",
          "services"
        ]
      },
      "UpdateQuoteDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "description": "Campo clientId",
            "example": "exemplo"
          },
          "discount": {
            "type": "number",
            "description": "Campo discount",
            "example": 1
          },
          "travelFee": {
            "type": "number",
            "description": "Campo travelFee",
            "example": 1
          },
          "materials": {
            "description": "Campo materials",
            "example": "exemplo",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "description": "Campo status",
            "example": "exemplo"
          },
          "services": {
            "description": "Campo services",
            "example": "exemplo",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "signature": {
            "type": "string",
            "description": "Campo signature",
            "example": "exemplo"
          }
        }
      },
...
      "UpdateServiceOrderDto": {
        "type": "object",
        "properties": {
          "technicianId": {
            "type": "string",
            "description": "Campo technicianId",
            "example": "exemplo"
          },
          "scheduledAt": {
            "type": "string",
            "description": "Campo scheduledAt",
            "example": "exemplo"
          },
          "totalValue": {
            "type": "number",
            "description": "Campo totalValue",
            "example": 1
          },
          "status": {
            "type": "string",
            "description": "Campo status",
            "example": "exemplo"
          },
          "observations": {
            "type": "string",
            "description": "Campo observations",
            "example": "exemplo"
          },
          "services": {
            "description": "Campo services",
            "example": "exemplo",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "materials": {
            "description": "Campo materials",
            "example": "exemplo",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}